I tried for a long while trying to implement a sliding window for finding patterns.
All of that work resulted in inefficiency. I also tried to get ChatGPT to help me figure it a non-janky pattern finder with a sliding window, but same thing.

So real changes are just getting ready the jailbreak patches. Just made empty functions, and I now know what some of the patches from sn0wbreeze are. That means the functions that provide patterns are mostly incorrect, except cs_enforcement and amfi::_memcmp.