Rework the entire offset finding.
Also, I know why it took a while for everything to work.

My kernel version string finder was looping after it already found a match. That is fixed now.

Also added a function I should made a while back, though I wanted to make one anyway, called crunchPatternUntilTrue().
This was very needed so it would decrease the amount of frustration in getting my offset finder to finally work with patterns that were off by a few bytes or so.

5.1.1 is for sure supported. But my new function I added will make manual pattern finding MUCH easier.